requirements.txt
Flask
torch>=2.2.0
Pillow
PyMuPDF
docx2txt
pandas
psutil
fastapi
uvicorn[standard]
python-multipart
aiofiles
transformers==4.52.4
easyocr
nltk
vllm>=0.8.3
accelerate>=0.26.0
sentencepiece
protobuf
litellm
json-repair
loguru
tenacity
xgrammar>=0.1.17
# DocExt dependencies
gradio>=5.23.0
mdpd
python-dotenv
python-levenshtein>=0.27.1
tabulate
types-requests